+++ docs/src/examples/components/Dropdown/Types/DropdownExample.shorthand.steps.ts
@@ -5,12 +5,19 @@ const selectors = {
   item: (itemIndex: number) => `.${Dropdown.slotClassNames.itemsList} li:nth-child(${itemIndex})`,
 }
 
-const steps = [
+const steps: ScreenerSteps = [
   steps => steps.click(selectors.triggerButton).snapshot('Shows list'),
-  steps => steps.click(selectors.item(3)).snapshot('Selects an item'),
-  steps => steps.click(selectors.triggerButton).snapshot('Opens with selected item highlighted'),
-  steps => steps.hover(selectors.item(2)).snapshot('Highlights another item'),
-  steps => steps.click(selectors.triggerButton).snapshot('Closes the list'),
+  steps =>
+    steps
+      .click(selectors.triggerButton)
+      .click(selectors.item(3))
+      .snapshot('Selects an item')
+      .click(selectors.triggerButton)
+      .snapshot('Opens with selected item highlighted')
+      .hover(selectors.item(2))
+      .snapshot('Highlights another item')
+      .click(selectors.triggerButton)
+      .snapshot('Closes the list'),
 ]
 
 export default steps